# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Gundam Platform application",
        "title": "Gundam Platform API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/auctions": {
            "get": {
                "description": "Retrieves upcoming and ongoing auctions from the platform.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auctions"
                ],
                "summary": "Get platform auctions",
                "parameters": [
                    {
                        "enum": [
                            "scheduled",
                            "active"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of auctions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.Auction"
                            }
                        }
                    }
                }
            }
        },
        "/auctions/{auctionID}": {
            "get": {
                "description": "Retrieves details of a specific auction by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auctions"
                ],
                "summary": "Get auction details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the auction",
                        "name": "auctionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the auction",
                        "schema": {
                            "$ref": "#/definitions/db.Auction"
                        }
                    }
                }
            }
        },
...
        "/users/me/auctions/{auctionID}/participate": {
            "post": {
                "security": [
                    {
                        "accessToken": []
                    }
                ],
                "description": "User deposits money to participate in an active auction.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auctions"
                ],
                "summary": "Participate in an auction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auction ID",
                        "name": "auctionID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.Auction": {
            "type": "object",
            "required": [
                "bid_increment",
                "buy_now_price",
                "canceled_by",
                "canceled_reason",
                "created_at",
                "current_price",
                "deposit_amount",
                "deposit_rate",
                "end_time",
                "gundam_id",
                "gundam_snapshot",
                "id",
                "order_id",
                "request_id",
                "seller_id",
                "start_time",
                "starting_price",
                "status",
                "total_bids",
                "total_participants",
                "updated_at",
                "winner_payment_deadline",
                "winning_bid_id"
            ],
            "properties": {
                "bid_increment": {
                    "type": "integer"
                },
                "buy_now_price": {
                    "type": "integer"
                },
                "canceled_by": {
                    "type": "string"
                },
                "canceled_reason": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_price": {
                    "type": "integer"
                },
                "deposit_amount": {
                    "type": "integer"
                },
                "deposit_rate": {
                    "type": "number"
                },
                "end_time": {
                    "type": "string"
                },
                "gundam_id": {
                    "type": "integer"
                },
                "gundam_snapshot": {
                    "$ref": "#/definitions/db.GundamSnapshot"
                },
                "id": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "seller_id": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "starting_price": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/db.AuctionStatus"
                },
                "total_bids": {
                    "type": "integer"
                },
                "total_participants": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "winner_payment_deadline": {
                    "type": "string"
                },
                "winning_bid_id": {
                    "type": "string"
                }
            }
        },
...
        "db.AuctionStatus": {
            "type": "string",
            "enum": [
                "scheduled",
                "active",
                "ended",
                "completed",
                "canceled",
                "failed"
            ],
            "x-enum-varnames": [
                "AuctionStatusScheduled",
                "AuctionStatusActive",
                "AuctionStatusEnded",
                "AuctionStatusCompleted",
                "AuctionStatusCanceled",
                "AuctionStatusFailed"
            ]
        },
...
        "db.GundamSnapshot": {
            "type": "object",
            "required": [
                "grade",
                "id",
                "image_url",
                "name",
                "quantity",
                "scale",
                "slug",
                "weight"
            ],
            "properties": {
                "grade": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "scale": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "accessToken": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}